@@ -45,14 +45,14 @@
   },
   "dependencies": {
     "@types/graphql": "^14.0.3",
-    "@types/jsonwebtoken": "<7.2.1",
+    "@types/jsonwebtoken": "^8.3.2",
     "@types/koa": "^2.0.44",
     "@types/pg": "^7.4.10",
     "@types/ws": "^6.0.1",
     "body-parser": "^1.15.2",
-    "chalk": "^1.1.3",
+    "chalk": "^2.4.2",
     "commander": "^2.19.0",
-    "debug": "^2.3.3",
+    "debug": "^4.1.1",
     "finalhandler": "^1.0.6",
     "graphql": "^0.6.0 || ^0.7.0 || ^0.8.0-b || ^0.9.0 || ^0.10.0 || ^0.11.0 || ^0.12.0 || ^0.13.0 || ^14.0.2",
     "http-errors": "^1.5.1",
@@ -61,7 +61,7 @@
     "lru-cache": ">=4 <5",
     "parseurl": "^1.3.2",
     "pg": ">=6.1.0 <8",
-    "pg-connection-string": "^0.1.3",
+    "pg-connection-string": "^2.0.0",
     "pg-sql2": "^2.2.1",
     "postgraphile-core": "4.4.0-beta.12",
     "subscriptions-transport-ws": "^0.9.15",
@@ -69,67 +69,56 @@
     "ws": "^6.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.0.0",
-    "@babel/plugin-transform-runtime": "7.0.0",
-    "@babel/preset-env": "7.0.0",
+    "@babel/core": "7.4.3",
+    "@babel/plugin-proposal-class-properties": "7.4.0",
+    "@babel/plugin-transform-runtime": "7.4.3",
+    "@babel/preset-env": "7.4.3",
     "@babel/preset-react": "7.0.0",
-    "@babel/runtime": "7.0.0",
-    "@types/chalk": "^0.4.30",
-    "@types/commander": ">=2.3.30 <2.12.0",
-    "@types/debug": "^0.0.29",
-    "@types/finalhandler": "^0.0.33",
+    "@babel/runtime": "7.4.3",
+    "@types/chalk": "^2.2.0",
+    "@types/commander": "^2.12.2",
+    "@types/debug": "^4.1.4",
+    "@types/finalhandler": "^1.1.0",
     "@types/http-errors": "^1.6.1",
-    "@types/jest": "^16.0.4",
-    "@types/lru-cache": "^4.1.1",
+    "@types/jest": "^24.0.11",
+    "@types/lru-cache": ">=4 <5",
     "@types/node": "^10.9.4",
     "@types/parseurl": "^1.3.1",
-    "babel-loader": "8.0.2",
+    "babel-loader": "8.0.5",
     "connect": "^3.5.0",
-    "css-loader": "1.0.0",
+    "css-loader": "2.1.1",
     "express": "^4.14.0",
-    "fastify": "^1.13.0",
+    "fastify": "^2.2.0",
     "html-webpack-inline-source-plugin": "^0.0.10",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "^18.1.0",
+    "jest": "^24.7.1",
     "koa": "^2.5.1",
     "koa-compress": "^3.0.0",
     "koa-mount": "^4.0.0",
     "nodemon": "^1.11.0",
-    "pg-minify": "~0.5.3",
+    "pg-minify": "~1.0.0",
     "prettier": "^1.15.2",
-    "source-map-support": "^0.4.6",
+    "source-map-support": "^0.5.12",
     "style-loader": "^0.23.0",
     "superagent": "^4.1.0",
-    "ts-node": "^2.0.0",
+    "ts-node": "^8.1.0",
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.14.0",
-    "typescript": "^2.9.2",
+    "typescript": "^3.4.3",
     "webpack": "^4.17.2"
   },
   "jest": {
     "transform": {
       ".*": "<rootDir>/resources/jest-preprocessor.js"
     },
-    "moduleFileExtensions": [
-      "ts",
-      "js",
-      "json"
-    ],
-    "setupFiles": [
-      "<rootDir>/resources/jest-setup.js"
-    ],
+    "moduleFileExtensions": ["ts", "js", "json"],
+    "setupFiles": ["<rootDir>/resources/jest-setup.js"],
     "browser": false,
     "testEnvironment": "node",
-    "testPathDirs": [
-      "<rootDir>/src"
-    ],
+    "roots": ["<rootDir>/src"],
     "testRegex": "/__tests__/[^.]+-test.(t|j)s$"
   },
-  "files": [
-    "sponsors.json",
-    "build"
-  ],
+  "files": ["sponsors.json", "build"],
   "engines": {
     "node": ">=8.6"
   }
